+++ lib/Controller/EditorController.php
@@ -681,9 +681,15 @@ class EditorController extends Controller {
 
         if ($file === null
             && !empty($link)) {
-            $fileId = $this->getFileIdByLink($link);
+            [$fileId, $redirect] = $this->getFileIdByLink($link);
             if (!empty($fileId)) {
                 list($file, $error, $share) = $this->getFile($userId, $fileId);
+            } elseif ($redirect) {
+                $response = [
+                    "url" => $link,
+                ];
+
+                return $response;
             }
         }
 
@@ -709,6 +715,11 @@ class EditorController extends Controller {
         ];
 
         if (!empty($this->config->getDocumentServerSecret())) {
+            $now = time();
+            $iat = $now;
+            $exp = $now + $this->config->getJwtExpiration() * 60;
+            $response["iat"] = $iat;
+            $response["exp"] = $exp;
             $token = \Firebase\JWT\JWT::encode($response, $this->config->getDocumentServerSecret(), "HS256");
             $response["token"] = $token;
         }
@@ -1117,6 +1128,11 @@ class EditorController extends Controller {
         }
 
         if (!empty($this->config->getDocumentServerSecret())) {
+            $now = time();
+            $iat = $now;
+            $exp = $now + $this->config->getJwtExpiration() * 60;
+            $result["iat"] = $iat;
+            $result["exp"] = $exp;
             $token = \Firebase\JWT\JWT::encode($result, $this->config->getDocumentServerSecret(), "HS256");
             $result["token"] = $token;
         }
@@ -1228,6 +1244,11 @@ class EditorController extends Controller {
         ];
 
         if (!empty($this->config->getDocumentServerSecret())) {
+            $now = time();
+            $iat = $now;
+            $exp = $now + $this->config->getJwtExpiration() * 60;
+            $result["iat"] = $iat;
+            $result["exp"] = $exp;
             $token = \Firebase\JWT\JWT::encode($result, $this->config->getDocumentServerSecret(), "HS256");
             $result["token"] = $token;
         }
@@ -1249,7 +1270,7 @@ class EditorController extends Controller {
     public function download($fileId, $toExtension = null, $template = false) {
         $this->logger->debug("Download: $fileId $toExtension");
 
-        if (!$this->config->isUserAllowedToUse()) {
+        if (!$this->config->isUserAllowedToUse() || $this->config->getDisableDownload()) {
             return $this->renderError($this->trans->t("Not permitted"));
         }
 
@@ -1571,31 +1592,26 @@ class EditorController extends Controller {
      *
      * @param string $link - link to the file
      *
-     * @return string|null
+     * @return array
      */
     private function getFileIdByLink(string $link) {
         $path = parse_url($link, PHP_URL_PATH);
         $encodedPath = array_map("urlencode", explode("/", $path));
-        $link = str_replace($path, implode("/", $encodedPath), $link);
-        if (filter_var($link, FILTER_VALIDATE_URL) === false) {
-            return null;
+        $parsedLink = str_replace($path, implode("/", $encodedPath), $link);
+        if (filter_var($parsedLink, FILTER_VALIDATE_URL) === false) {
+            return [null, true];
         }
 
-        if (!empty($this->config->getStorageUrl())) {
-            $storageUrl = $this->config->getStorageUrl();
-        } else {
-            $storageUrl = $this->urlGenerator->getAbsoluteURL("/");
+        $storageUrl = $this->urlGenerator->getAbsoluteURL("/");
+        if (parse_url($parsedLink, PHP_URL_HOST) !== parse_url($storageUrl, PHP_URL_HOST)) {
+            return [null, true];
         }
 
-        if (parse_url($link, PHP_URL_HOST) !== parse_url($storageUrl, PHP_URL_HOST)) {
-            return null;
+        if (preg_match('/\/(files|f|onlyoffice)\/(\d+)/', $parsedLink, $matches)) {
+            return [$matches[2], false];
         }
 
-        if (preg_match('/\/(files|f|onlyoffice)\/(\d+)/', $link, $matches)) {
-            return $matches[2];
-        }
-
-        return null;
+        return [null, false];
     }
 
     /**